+++ modules/xls_export/lib/open_project/xls_export/xls_views/cost_entry_table.xls.rb
@@ -1,6 +1,50 @@
-require_relative './one_dimensional_table.xls'
+class OpenProject::XlsExport::XlsViews::CostEntryTable < OpenProject::XlsExport::XlsViews
+  def generate
+    @spreadsheet = OpenProject::XlsExport::SpreadsheetBuilder.new(I18n.t(:label_money))
+    default_query = serialize_query_without_hidden(@query)
 
-class OpenProject::XlsExport::XlsViews::CostEntryTable < OpenProject::XlsExport::XlsViews::OneDimensionalTable
+    available_cost_type_tabs(options[:cost_types]).each_with_index do |(unit_id, name), idx|
+      setup_query_for_tab(default_query, unit_id)
+
+      spreadsheet.worksheet(idx, name)
+      build_spreadsheet
+    end
+
+    spreadsheet
+  end
+
+  def setup_query_for_tab(query, unit_id)
+    @query = CostQuery.deserialize(query)
+    @cost_type = nil
+    @unit_id = unit_id
+
+    if @unit_id != 0
+      @query.filter :cost_type_id, operator: '=', value: @unit_id.to_s
+      @cost_type = CostType.find(unit_id) if unit_id.positive?
+    end
+  end
+
+  def build_spreadsheet
+    set_title
+
+    build_header
+    format_columns
+    build_cost_rows
+    build_footer
+
+    spreadsheet
+  end
+
+  def build_header
+    spreadsheet.add_headers(headers)
+  end
+
+  def build_cost_rows
+    query.each_direct_result do |result|
+      spreadsheet.add_row(cost_row(result))
+    end
+  end
+
   def format_columns
     spreadsheet.add_format_option_to_column(headers.length - 3,
                                             number_format: number_format)
@@ -25,11 +69,19 @@ class OpenProject::XlsExport::XlsViews::CostEntryTable < OpenProject::XlsExport:
   def build_footer
     footer = [''] * cost_entry_attributes.size
     footer += if show_result(query, 0) != show_result(query)
-                [show_result(query), '', show_result(query, 0)]
+                one_unit_type_footer
               else
-                ['', '', show_result(query)]
+                multiple_unit_types_footer
               end
     spreadsheet.add_sums(footer) # footer
+  end
+
+  def one_unit_type_footer
+    [show_result(query), '', show_result(query, 0)]
+  end
+
+  def multiple_unit_types_footer
+    ['', '', show_result(query)]
   end
 
   def headers
